+++ tests/test_models.py
@@ -13,7 +13,7 @@ from tot.evaluation.metrics import ERROR_FUNCTIONS
 from tot.models.models_naive import NaiveModel, SeasonalNaiveModel
 from tot.models.models_neuralprophet import NeuralProphetModel, TorchProphetModel
 from tot.models.models_prophet import ProphetModel
-from tot.models.models_simple import LinearRegressionModel, RandomForestModel, DartsForecastingModel
+from tot.models.models_darts import LinearRegressionModel, RandomForestModel, DartsForecastingModel
 
 log = logging.getLogger("tot.test")
 log.setLevel("WARNING")
