+++ tests/test_processing_panel_data.py
@@ -6,13 +6,14 @@ import pathlib
 
 import pandas as pd
 import pytest
+from darts.models import NaiveDrift
 
 from tot.benchmark import CrossValidationBenchmark, ManualCVBenchmark, SimpleBenchmark
 from tot.datasets.dataset import Dataset
 from tot.experiment import CrossValidationExperiment
 from tot.models.models_naive import NaiveModel, SeasonalNaiveModel
 from tot.models.models_neuralprophet import NeuralProphetModel, TorchProphetModel
-from tot.models.models_simple import LinearRegressionModel
+from tot.models.models_darts import LinearRegressionModel, RandomForestModel, DartsForecastingModel
 
 log = logging.getLogger("tot.test")
 log.setLevel("WARNING")
@@ -85,6 +86,9 @@ def test_benchmark_panel_data_input():
         (LinearRegressionModel, {"lags": 24, "output_chunk_length": 8, "n_forecasts": 8}),
         (NaiveModel, {"n_forecasts": 8}),
         (SeasonalNaiveModel, {"n_forecasts": 8, "season_length": 24}),
+        (RandomForestModel, {"lags": 24, "output_chunk_length": 8, "n_forecasts": 8}),
+        (DartsForecastingModel,
+         {"model": NaiveDrift, "retrain": True, "lags": 12, "n_forecasts": 4})
     ]
     log.debug("{}".format(model_classes_and_params))
 
@@ -170,7 +174,7 @@ def test_benchmark_manualCV_global_modeling():
     slice_idx = 0
     log.info("Creating a date intersection between df1 and df2")
     for df_name in ["df1", "df2"]:
-        df_aux = peyton_manning_df_aux.iloc[slice_idx : slice_idx + 100]
+        df_aux = peyton_manning_df_aux.iloc[slice_idx: slice_idx + 100]
         df_aux = df_aux.assign(ID=df_name)
         peyton_manning_df = pd.concat((peyton_manning_df, df_aux), ignore_index=True)
         slice_idx = slice_idx + 100
