+++ tot/models/__init__.py
@@ -2,6 +2,6 @@
 from .models_naive import NaiveModel, SeasonalNaiveModel  # noqa: F401 to evade flake8
 from .models_neuralprophet import NeuralProphetModel, TorchProphetModel  # noqa: F401 to evade flake8
 from .models_prophet import ProphetModel  # noqa: F401 to evade flake8
-from .models_simple import DartsForecastingModel  # noqa: F401 to evade flake8
-from .models_simple import LinearRegressionModel  # noqa: F401 to evade flake8
-from .models_simple import RandomForestModel  # noqa: F401 to evade flake8
+from .models_darts import DartsForecastingModel  # noqa: F401 to evade flake8
+from .models_darts import LinearRegressionModel  # noqa: F401 to evade flake8
+from .models_darts import RandomForestModel  # noqa: F401 to evade flake8
